requirements-test.txt
coveralls
django-environ
mysqlclient
psycopg2
sphinx
sphinx_rtd_theme
freezegun
django-extensions
django-rest-auth>=0.9.3,<0.12.0
django-allauth>=0.41.0,<0.42.0
openwisp-utils[qa]>=0.4.0
